requirements.txt
python-dotenv
playwright
langchain
langgraph
openai
Pillow
